+++ zeronet-protocol/lib/index.js
@@ -1,7 +1,13 @@
 "use strict"
 
-function Protocol() {
+const protobuf = require("protocol-buffers")
 
+function Protocol() {
+  const self = this
+  const protos = self.protos = {}
+  self.handle = (name, opt, handler) => {
+
+  }
 }
 
 module.exports = Protocol
