+++ zeronet-protocol/lib/zero/peer-request-handler.js
